requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
pydantic
pydantic-settings
python-dotenv
httpx
apscheduler
python-dateutil